+++ src/components/common/Navbar.jsx
@@ -1,21 +1,27 @@
 import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
+import axios from "axios";
 
 import { Menu, MenuItem, Dialog, DialogContent } from "@mui/material";
-import KakaoSocialLogin from "./KakaoSocialLogin";
 
 import styled from "styled-components";
 
 import profile from "../../assets/icons/profile.png";
 import register from "../../assets/icons/plus.png";
+import kakao from "../../assets/icons/kakao.png";
+
+import dotenv from "dotenv";
+dotenv.config();
 
 const Navbar = () => {
 	const [isLoggedIn, setLoggedIn] = useState(false);
 	const [modalOpen, setModalOpen] = useState(false);
 	const [anchorEl, setAnchorEl] = useState(null);
 	const open = Boolean(anchorEl);
-
 	const userToken = localStorage.getItem("token");
+	const navigate = useNavigate();
+
+	const { Kakao } = window;
 
 	useEffect(() => {
 		if (userToken) setLoggedIn(true);
@@ -37,6 +43,42 @@ const Navbar = () => {
 		setModalOpen(false);
 	};
 
+	const kakaoLoginHandler = () => {
+		Kakao.Auth.login({
+			success: (response) => {
+				const token = response.access_token;
+				console.log(token);
+				axios
+					.post(
+						"https://api.fevertime.shop/login/kakao",
+						JSON.stringify({ token: token }),
+						{
+							headers: { "Content-Type": "application/json" },
+						}
+					)
+					.then((res) => {
+						console.log(res);
+						if (res.data.token) {
+							localStorage.setItem("token", res.data.token);
+							handleModalClose();
+							navigate(0);
+						}
+					})
+					.catch((err) => console.error(err));
+			},
+			fail: (err) => {
+				console.log(err);
+			},
+		});
+	};
+
+	const kakaoLogoutHandler = () => {
+		Kakao.Auth.logout(() => {
+			localStorage.removeItem("token");
+			navigate(0);
+		});
+	};
+
 	return (
 		<>
 			{!isLoggedIn ? (
@@ -52,10 +94,12 @@ const Navbar = () => {
 							<StyledDialogContent>
 								삼세번과 함께 <br />
 								건강한 습관을 만들어보세요!
-								<KakaoSocialLogin
-									onClose={handleModalClose}
-									setLoggedIn={setLoggedIn}
-								/>
+								<KakaoLoginButton onClick={kakaoLoginHandler}>
+									<LoginIcon>
+										<img src={kakao} alt="Kakao icon" />
+									</LoginIcon>
+									<LoginText>카카오 로그인</LoginText>
+								</KakaoLoginButton>
 							</StyledDialogContent>
 						</StyledDialog>
 					</NavContainer>
@@ -80,8 +124,10 @@ const Navbar = () => {
 								open={open}
 								onClose={handleDropDownClose}
 							>
-								<MenuItem>My page</MenuItem>
-								<MenuItem>Logout</MenuItem>
+								<NavLink to="/my">
+									<MenuItem>My page</MenuItem>
+								</NavLink>
+								<MenuItem onClick={kakaoLogoutHandler}>Logout</MenuItem>
 							</StyledMenu>
 							<NavRegister>
 								<img src={register} alt="Register icon" />
@@ -178,4 +224,31 @@ const StyledDialogContent = styled(DialogContent)`
 	display: flex;
 	flex-direction: column;
 	gap: 4rem;
+`;
+
+const KakaoLoginButton = styled.button`
+	width: 31.7rem;
+	height: 4rem;
+	padding: 1rem;
+	font-size: 1.6rem;
+	font-weight: bold;
+	letter-spacing: 0.2rem;
+	background-color: #ffe812;
+	border: none;
+	border-radius: 0.5rem;
+	outline: none;
+	cursor: pointer;
+	display: flex;
+	align-items: center;
+	&:hover {
+		opacity: 0.7;
+	}
+`;
+
+const LoginIcon = styled.span`
+	flex: 0.1;
+`;
+
+const LoginText = styled.span`
+	flex: 0.8;
 `;
